requirements.txt
streamlit==1.31.0
requests==2.31.0
pyyaml==6.0.1
